+++ prepare_project_gke.bash
@@ -94,6 +94,8 @@ echo -e ${GREEN?}'Granting the required permissions.'${RESET?}
 gcloud projects add-iam-policy-binding -q ${PROJECT?} \
   --member serviceAccount:${PROJECT_NUMBER?}-compute@developer.gserviceaccount.com --role roles/cloudkms.cryptoKeyEncrypterDecrypter
 gcloud projects add-iam-policy-binding -q ${PROJECT?} \
+  --member serviceAccount:${PROJECT_NUMBER?}-compute@developer.gserviceaccount.com --role roles/cloudkms.signerVerifier
+gcloud projects add-iam-policy-binding -q ${PROJECT?} \
   --member serviceAccount:${PROJECT_NUMBER?}-compute@developer.gserviceaccount.com --role roles/iam.serviceAccountTokenCreator
 gcloud projects add-iam-policy-binding -q ${PROJECT?} \
   --member serviceAccount:${PROJECT_NUMBER?}-compute@developer.gserviceaccount.com --role roles/cloudsql.client